Move/migrate utils.js functions into .ts utils directory
