+++ metaspace/graphql/esConnector.ts
@@ -5,8 +5,8 @@ import {UserProjectRoleOptions as UPRO} from './src/modules/project/model';
 import * as elasticsearch from 'elasticsearch';
 import * as sprintf from 'sprintf-js';
 import config from './src/utils/config';
+import logger from './src/utils/logger';
 import {datasetFilters} from './datasetFilters.js';
-import {logger} from './utils';
 import {ContextUser, UserProjectRoles} from './src/context';
 import {AnnotationFilter, AnnotationOrderBy, DatasetFilter, DatasetOrderBy, SortingOrder} from './src/binding';
