+++ metaspace/graphql/src/modules/dataset/controller/Subscription.ts
@@ -1,7 +1,8 @@
 import * as Amqplib from 'amqplib';
 import * as Sentry from '@sentry/node';
 import {esDatasetByID} from '../../../../esConnector';
-import {logger, wait} from '../../../../utils';
+import logger from '../../../utils/logger';
+import wait from '../../../utils/wait';
 import config from '../../../utils/config';
 import {DatasetStatus} from '../model';
 import {Context} from '../../../context';
