+++ metaspace/graphql/src/modules/group/controller.ts
@@ -7,10 +7,11 @@ import {Dataset as DatasetModel} from '../dataset/model';
 import {Group, UserGroup, UserGroupRole} from '../../binding';
 import {Context, ContextUser} from '../../context';
 import {Scope, ScopeRoleOptions, UserGroupSource} from '../../bindingTypes';
-import {findUserByEmail, logger, LooselyCompatible} from '../../utils';
+import {findUserByEmail, LooselyCompatible} from '../../utils';
 import {sendInvitationEmail} from '../auth';
 import {sendAcceptanceEmail, sentGroupOrProjectInvitationEmail, sendRequestAccessEmail} from '../groupOrProject/email';
 import config from '../../utils/config';
+import logger from '../../utils/logger';
 import {createInactiveUser} from '../auth/operation';
 import {smAPIUpdateDataset} from '../../utils/smAPI';
 import {getDatasetForEditing} from '../dataset/operation/getDatasetForEditing';
