+++ metaspace/graphql/src/modules/user/controller.ts
@@ -11,7 +11,8 @@ import {UserProject as UserProjectModel} from '../project/model';
 import {Context, ContextUser} from '../../context';
 import {ScopeRole, ScopeRoleOptions as SRO, UserProjectSource, UserSource} from '../../bindingTypes';
 import {sendEmailVerificationToken} from '../auth/operation';
-import {logger, LooselyCompatible} from '../../utils';
+import {LooselyCompatible} from '../../utils';
+import logger from '../../utils/logger';
 import {convertUserToUserSource} from './util/convertUserToUserSource';
 import {smAPIUpdateDataset} from '../../utils/smAPI';
 import {deleteDataset} from '../dataset/operation/deleteDataset';
