+++ metaspace/graphql/src/utils/wait.ts
@@ -1,0 +1,3 @@
+export default async function wait(ms: number) {
+  return new Promise(resolve => setTimeout(resolve, ms));
+}
